Testing merging into a single file
